lockfree: fix test case on certain compilers
git-svn-id: http://svn.boost.org/svn/boost/trunk@81741 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/lockfree/test/test_common.hpp
@@ -111,7 +111,7 @@ struct queue_stress_tester
 
         cout << "reader threads joined" << endl;
 
-        BOOST_REQUIRE_EQUAL(data.count_nodes(), 0);
+        BOOST_REQUIRE_EQUAL(data.count_nodes(), (size_t)0);
         BOOST_REQUIRE(stk.empty());
 
         BOOST_REQUIRE_EQUAL(push_count, pop_count);
